Refactor code structure and remove redundant sections for improved readability and maintainability

+++ apps/services/projects/src/services/deployments.service.ts
@@ -1,9 +1,10 @@
-import { createAsyncErrHandler, createGrpcErrorHandler, GrpcResponse } from "@shipoff/services-commons";
+import { createAsyncErrHandler, createGrpcErrorHandler, GrpcAppError, GrpcResponse } from "@shipoff/services-commons";
 import { Database, dbService } from "@/db/db-service";
 import authExternalService, { AuthExternalService } from "@/externals/auth.external.service";
 import { DeleteDeploymentRequestBodyType, GetAllDeploymentsRequestBodyType, GetDeploymentRequestBodyType, RedeployRequestBodyType } from "@/types/deployments";
 import { DeploymentEventProducerService } from "@/producer/deployment.producer";
 import { logger } from "@/libs/winston";
+import { status } from "@grpc/grpc-js";
 
 
 export class DeploymentsService {
@@ -12,6 +13,7 @@ export class DeploymentsService {
     private _errHandler : ReturnType<typeof createGrpcErrorHandler>
     private _asyncErrHandler: ReturnType<typeof createAsyncErrHandler>  
     private _deploymentProducer : DeploymentEventProducerService;
+    private _deletableStatuses = ["FAILED","INACTIVE"];
 
     constructor() {
         this._errHandler = createGrpcErrorHandler({subServiceName:"DEPLOYMENT_SERVICE",logger});
@@ -79,7 +81,9 @@ export class DeploymentsService {
                 resourceId: projectId,
                 errMsg: "You do not have permission to delete this deployment"
             });
-            const deployment = await this._dbService.deleteDeploymentById(deploymentId);
+            const deployment = await this._dbService.findUniqueDeploymentById(deploymentId);
+            if(!this._deletableStatuses.includes(deployment.status))throw new GrpcAppError(status.FAILED_PRECONDITION,"Deployment is not in a deletable state, only deployments in FAILED or INACTIVE state can be deleted");
+            const deletedDeployment = await this._dbService.deleteDeploymentById(deploymentId);
             this._asyncErrHandler.call(this._deploymentProducer.publishDeploymentRequested({
                 event: "DELETED",
                 projectId,
@@ -89,7 +93,7 @@ export class DeploymentsService {
                 commitHash:deployment.commitHash,
                 requestId:reqMeta.requestId
             }),"DELETE-DEPLOYMENT",reqMeta.requestId);
-            return GrpcResponse.OK({}, "Deployment deleted");
+            return GrpcResponse.OK(deletedDeployment, "Deployment deleted");
         } catch (e:any) {
             return this._errHandler(e, "DELETE-DEPLOYMENT",reqMeta.requestId);
         }
@@ -104,7 +108,9 @@ export class DeploymentsService {
                 resourceId: projectId,
                 errMsg: "You do not have permission to redeploy this deployment"
             });
-            const deployment = await this._dbService.findUniqueDeploymentById(deploymentId);
+            const deployment = await this._dbService.updateDeploymentById(deploymentId,projectId,{
+                status:"QUEUED"
+            })
             this._asyncErrHandler.call(this._deploymentProducer.publishDeploymentRequested({
                 event:"CREATED",
                 projectId,
